pyproject.toml
[project]
name = "nexus"
version = "2.0.0"
description = "Self-hosted homelab solution"
authors = [
    {name = "Ryan Liu", email = "ryanliu6@pm.me"},
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click>=8.1.2",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "discord.py>=2.3.0",
    "aiohttp>=3.9.0",
    "ansible>=8.0.0",
    "ansible-vault>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0", 
]

[project.scripts]
nexus-deploy = "scripts.deploy:main"
nexus-generate-dashboard = "scripts.generate_dashboard:main"
nexus-restore = "scripts.restore:main"
nexus-health = "scripts.health_check:main"
nexus-operations = "scripts.operations:main"
nexus-alert-bot = "scripts.alert_bot:main"

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexus"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.format]
quote-style = "double"
line-ending = "lf"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "B",  # flake8-bugbear
    "I",  # isort
    "UP", # pyupgrade
    "RUF", # Ruff-specific rules
    "D",   # pydocstyle (Added this because user wants docstrings)
]
ignore = [
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D107",  # Missing docstring in __init__
    "D203",  # 1 blank line required before class docstring
    "D213",  # Multi-line docstring summary should start at the second line
    "D406",  # Section name should end with a newline
    "D407",  # Missing dashed underline after section
    "D413",  # Missing blank line after last section
    "COM812",  # Missing trailing comma
    "ISC001",  # Single line import statements
    "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
    "UP007",  # Use X | Y for type annotations (we prefer Optional for readability)
    "UP045",  # Use X | None instead of Optional[X] (we prefer Optional for readability)
]

exclude = [
    ".*/tests/.*",
    'conftest\.py',
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.mypy]
python_version = "3.9"
mypy_path = "src"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = [
    ".*/tests/.*",
    'conftest\.py',
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov"
testpaths = [
    "tests",
]